+++ CHANGELOG.rst
@@ -105,6 +105,9 @@ Additions
     DependNode.getAliasAttr, DependNode.hasAttribute, DependNode.isNewAttribute,
     DependNode.setFlag, ObjectSet.isMember
 
+- DependNode.deleteAttr now accepts Attribute objects (first verifying that it
+  is an attribute on the same node as self)
+
 ----------------------------------
 Bug Fixes
 ----------------------------------